- split views.py
